the create page with 3 categories all ready, resolve the map issue, fix the broken drawer
